SGT-1191: Numeric overflow (int/long) caused IllegalArgumentException because the file position was calculated as negative, now fixed. The problem could lead to incorrect reading of FSFS repositories of format 7 (i.e. created with Subversion 1.9).

+++ svnkit/src/main/java/org/tmatesoft/svn/core/internal/io/fs/index/FSPackedNumbersStream.java
@@ -114,11 +114,11 @@ public class FSPackedNumbersStream {
                 int j = i;
                 long shift = 0;
                 while ((bytes[j] & 0x80) != 0) {
-                    value += (bytes[j] & 0x7f) << shift;
+                    value += (long)(bytes[j] & 0x7f) << shift;
                     shift += 7;
                     j++;
                 }
-                value += (bytes[j] & 0x7f) << shift;
+                value += (long)(bytes[j] & 0x7f) << shift;
                 int bytesProcessed = j - i + 1;
                 length = bytesProcessed;
                 i += bytesProcessed;
